+++ NEWS.rst
@@ -8,6 +8,11 @@ Latest (minor)
 --------------
 * Adding the shallow storage decoder which allows decoding objects larger
   than a single block or generation.
+* Adding new example of an on-the-fly encoder. This example shows how to
+  use an encoder before all symbols for that encoder has become available.
+  This can be very useful in cases where data is produced on-the-fly, e.g.
+  some form of streaming. The example can be found in the "encode_on_the_fly"
+  folder.
 
 8.0.0
 -----